Split implementation into multiple files, organize imports and add documentation
