+++ osu.Framework/Graphics/Visualisation/DrawableInspector.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Linq;
+
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Extensions;
